requirements.txt
streamlit
exa_py
readability-lxml
beautifulsoup4
rank-bm25
requests